proyectosCapacitación/primerProyecto/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
